+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Change Log
 
-## 1.0.0
-- Initial release
+## 1.0.2
+
+- Uses status bar for information rather than information messages.
+- No delay between switching and the confirmation message.
 
 ## 1.0.1
+
 - Minor documentation tweaks
+
+## 1.0.0
+
+- Initial release